:truck: Moved AzureSerialisableModel to serialisers
