+++ data_safe_haven/config/config.py
@@ -12,6 +12,7 @@ from pydantic import (
 
 from data_safe_haven.exceptions import DataSafeHavenConfigError
 from data_safe_haven.functions.validators import validate_unique_list
+from data_safe_haven.serialisers import AzureSerialisableModel
 from data_safe_haven.utility import (
     DatabaseSystem,
     LoggingSingleton,
@@ -26,8 +27,6 @@ from data_safe_haven.utility.annotated_types import (
     TimeZone,
     UniqueList,
 )
-
-from .azure_serialisable_model import AzureSerialisableModel
 
 
 class ConfigSectionAzure(BaseModel, validate_assignment=True):
