+++ data_safe_haven/config/pulumi.py
@@ -4,7 +4,7 @@ from typing import Any, ClassVar
 
 from pydantic import BaseModel
 
-from .azure_serialisable_model import AzureSerialisableModel
+from data_safe_haven.serialisers import AzureSerialisableModel
 
 
 class DSHPulumiProject(BaseModel, validate_assignment=True):
