+++ data_safe_haven/serialisers/__init__.py
@@ -1,5 +1,7 @@
+from .azure_serialisable_model import AzureSerialisableModel
 from .yaml_serialisable_model import YAMLSerialisableModel
 
 __all__ = [
+    "AzureSerialisableModel",
     "YAMLSerialisableModel",
 ]
