+++ data_safe_haven/serialisers/azure_serialisable_model.py
@@ -2,9 +2,10 @@
 
 from typing import Any, ClassVar, TypeVar
 
-from data_safe_haven.context import Context
+from data_safe_haven.context.context import Context
 from data_safe_haven.external import AzureApi
-from data_safe_haven.serialisers import YAMLSerialisableModel
+
+from .yaml_serialisable_model import YAMLSerialisableModel
 
 T = TypeVar("T", bound="AzureSerialisableModel")
